KNL-1233 removed all direct use of ehcaches in kernel, EXCEPT in SakaiCacheProvider.java (which is problematic and probably will need to stay for now) Added the cache settings for these caches to the kernel.properties so the original settings are retained without the ehcache specific beans
git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@308232 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ kernel/kernel-impl/src/main/java/org/sakaiproject/user/impl/AuthnCacheWatcher.java
@@ -21,19 +21,20 @@
 
 package org.sakaiproject.user.impl;
 
-import java.util.Observable;
-import java.util.Observer;
-
-import net.sf.ehcache.Cache;
-
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.sakaiproject.component.cover.ComponentManager;
 import org.sakaiproject.entity.api.EntityManager;
 import org.sakaiproject.entity.api.Reference;
 import org.sakaiproject.event.api.Event;
 import org.sakaiproject.event.api.EventTrackingService;
+import org.sakaiproject.memory.api.Cache;
+import org.sakaiproject.memory.api.MemoryService;
 import org.sakaiproject.user.api.UserDirectoryService;
 import org.sakaiproject.user.api.UserNotDefinedException;
+
+import java.util.Observable;
+import java.util.Observer;
 
 /**
  * This observer watches for user.add and user.upd events to invalidate the Authn cache
@@ -44,20 +45,20 @@ import org.sakaiproject.user.api.UserNotDefinedException;
 public class AuthnCacheWatcher implements Observer {
 
 	private static final Log log = LogFactory.getLog(AuthnCacheWatcher.class);
-	
-	private AuthenticationCache authenticationCache;
-	
-	private UserDirectoryService userDirectoryService;
-	
-	private EventTrackingService eventTrackingService;
-	
-	private EntityManager entityManager;
-	
 	//Copied from DbUserService as they are private
 	private static final String EIDCACHE = "eid:";
 	private static final String IDCACHE = "id:";
-	
+	private AuthenticationCache authenticationCache;
+	private UserDirectoryService userDirectoryService;
+	private EventTrackingService eventTrackingService;
+	private EntityManager entityManager;
+    private MemoryService memoryService;
 	private Cache userCache = null;
+	
+    public void setMemoryService(MemoryService memoryService) {
+        this.memoryService = memoryService;
+    }
+
 		public void setUserCache(Cache userCache) {
 		this.userCache = userCache;
 	}
@@ -85,11 +86,16 @@ public class AuthnCacheWatcher implements Observer {
 
 	public void init() {
 		log.info("init()");
+        if (userCache == null) { // this is the user id->eid mapping cache
+            userCache = memoryService.getCache("org.sakaiproject.user.api.UserDirectoryService");
+        }
 		eventTrackingService.addObserver(this);
 	}
 	
 	public void destroy() {
-		eventTrackingService.deleteObserver(this);
+        if (!ComponentManager.hasBeenClosed()) {
+            eventTrackingService.deleteObserver(this);
+        }
 	}
 	
 	public void update(Observable arg0, Object arg) {
